# static/api-docs/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "prose-pod-api",
		"description": "Prose Pod API server.",
		"contact": {
			"name": "Valerian Saliou",
			"email": "valerian@valeriansaliou.name"
		},
		"license": {
			"name": "MPL-2.0"
		},
		"version": "0.2.0"
	},
	"paths": {
		"/v1/init/first-account": {
			"put": {
				"tags": ["Init"],
				"summary": "Initialize the the first XMPP account",
				"description": "Initializes the first XMPP account on the server. It will be given the `ADMIN` role and this user will have to log in before making subsequent requests (like inviting other users). After one account is created, this unauthenticated route will always return `409 Conflict` to prevent unauthenticated users from creating admin accounts.",
				"operationId": "init_first_account",
				"security": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/InitFirstAccountRequest" }
						}
					}
				},
				"responses": {
					"201": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Member" }
							}
						},
						"headers": {
							"Location": { "$ref": "#/components/headers/Location" }
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"409": { "$ref": "#/components/responses/FirstAccountAlreadyCreated" }
				}
			}
		},
...
		"/v1/invitations": {
			"get": {
				"tags": ["Invitations"],
				"summary": "Get all invitations",
				"description": "Get all workspace invitations in a paginated manner.",
				"operationId": "get_invitations",
				"security": [
					{ "BearerAuth": [] }
				],
				"parameters": [
					{ "$ref": "#/components/parameters/PageNumber" },
					{ "$ref": "#/components/parameters/PageSize" },
					{ "$ref": "#/components/parameters/Until" }
				],
				"responses": {
					"200": {
						"description": "End of results",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": { "$ref": "#/components/schemas/WorkspaceInvitation" }
								}
							}
						},
						"headers": {
							"Pagination-Current-Page": { "$ref": "#/components/headers/PaginationCurrentPage" },
							"Pagination-Page-Size": { "$ref": "#/components/headers/PaginationPageSize" },
							"Pagination-Page-Count": { "$ref": "#/components/headers/PaginationPageCount" },
							"Pagination-Item-Count": { "$ref": "#/components/headers/PaginationItemCount" }
						}
					},
					"206": {
						"description": "Partial results",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": { "$ref": "#/components/schemas/WorkspaceInvitation" }
								}
							}
						},
						"headers": {
							"Pagination-Current-Page": { "$ref": "#/components/headers/PaginationCurrentPage" },
							"Pagination-Page-Size": { "$ref": "#/components/headers/PaginationPageSize" },
							"Pagination-Page-Count": { "$ref": "#/components/headers/PaginationPageCount" },
							"Pagination-Item-Count": { "$ref": "#/components/headers/PaginationItemCount" }
						}
					},
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			},
			"post": {
				"tags": ["Invitations"],
				"summary": "Invite a member",
				"description": "Invite a new member to the workspace.\n\nThe person will receive a notification with a link to join the workspace and create their account.",
				"operationId": "invite_member",
				"security": [
					{ "BearerAuth": [] }
				],
				"requestBody": {
					"description": "Information about who to invite.",
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/InviteMemberRequest" }
						}
					}
				},
				"responses": {
					"201": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/WorkspaceInvitation" }
							}
						},
						"headers": {
							"Location": { "$ref": "#/components/headers/Location" }
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/invitations/{invitationIdOrToken}": {
			"get": {
				"tags": ["Invitations"],
				"summary": "Get an invitation",
				"description": "Get information about a workspace invitation.\n\nIf an accept or reject token is used in the path and `token_type` is provided as a query parameter, will search for an invitation with the given token. This route can be used by the Prose Pod Dashboard to display some information about the invitation when someone opens an invitation link (like their predefined JID or role).",
				"operationId": "get_invitation",
				"security": [
					{ "BearerAuth": [] }
				],
				"parameters": [
					{ "$ref": "#/components/parameters/InvitationIdOrToken" },
					{
						"description": "Type of invitation token",
						"in": "query",
						"name": "token_type",
						"schema": { "$ref": "#/components/schemas/InvitationTokenType" },
						"required": false
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/WorkspaceInvitation" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/invitations/{invitationId}": {
			"delete": {
				"tags": ["Invitations"],
				"summary": "Cancel an invitation",
				"description": "Cancel a workspace invitation.",
				"operationId": "invitation_cancel",
				"security": [
					{ "BearerAuth": [] }
				],
				"parameters": [
					{ "$ref": "#/components/parameters/InvitationId" }
				],
				"responses": {
					"204": { "$ref": "#/components/responses/NoContent" },
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
		"/v1/invitations/{token}/accept": {
			"put": {
				"tags": ["Invitations"],
				"summary": "Accept an invitation",
				"description": "Accept a workspace invitation.",
				"operationId": "invitation_accept",
				"security": [],
				"parameters": [
					{ "$ref": "#/components/parameters/InvitationAcceptToken" }
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/AcceptWorkspaceInvitationRequest" }
						}
					}
				},
				"responses": {
					"204": { "$ref": "#/components/responses/NoContent" },
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"410": { "$ref": "#/components/responses/Gone" }
				}
			}
		},
		"/v1/invitations/{token}/reject": {
			"put": {
				"tags": ["Invitations"],
				"summary": "Reject an invitation",
				"description": "Reject a workspace invitation.",
				"operationId": "invitation_reject",
				"security": [],
				"parameters": [
					{ "$ref": "#/components/parameters/InvitationId" }
				],
				"responses": {
					"204": { "$ref": "#/components/responses/NoContent" },
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"410": { "$ref": "#/components/responses/Gone" }
				}
			}
		},
		"/v1/invitations/{invitationId}/resend": {
			"post": {
				"tags": ["Invitations"],
				"summary": "Resend an invitation",
				"description": "Resend a workspace invitation.",
				"operationId": "invitation_resend",
				"security": [
					{ "BearerAuth": [] }
				],
				"parameters": [
					{ "$ref": "#/components/parameters/InvitationId" }
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/WorkspaceInvitation" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
...
		"/v1/members/{memberId}": {
			"get": {
				"tags": ["Members"],
				"summary": "Get a member",
				"description": "Get information about one member.",
				"operationId": "get_member",
				"security": [
					{ "BearerAuth": [] }
				],
				"parameters": [
					{ "$ref": "#/components/parameters/MemberId" }
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Member" }
							}
						}
					},
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" },
					"404": { "$ref": "#/components/responses/NotFound" }
				}
			}
		},
...
		"/v1/server/config": {
			"put": {
				"tags": ["Init"],
				"summary": "Initialize the XMPP server",
				"description": "Initializes the XMPP server and returns its default configuration.",
				"operationId": "init_server_config",
				"security": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/InitServerConfigRequest" }
						}
					}
				},
				"responses": {
					"201": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						},
						"headers": {
							"Location": { "$ref": "#/components/headers/Location" }
						}
					},
					"409": { "$ref": "#/components/responses/ServerConfigAlreadyInitialized" }
				}
			},
			"get": {
				"tags": ["Server / Configuration"],
				"summary": "Get server config",
				"description": "Get the current configuration of the XMPP server.",
				"operationId": "get_server_config",
				"security": [
					{ "BearerAuth": [] }
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ServerConfig" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ServerConfigNotInitialized" },
					"401": { "$ref": "#/components/responses/Unauthorized" },
					"403": { "$ref": "#/components/responses/Forbidden" }
				}
			}
		},
...
		"/v1/workspace": {
			"put": {
				"tags": ["Init"],
				"summary": "Initialize the workspace",
				"description": "Initializes the workspace and returns its default configuration.",
				"operationId": "init_workspace",
				"security": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/InitWorkspaceRequest" }
						}
					}
				},
				"responses": {
					"201": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Workspace" }
							}
						},
						"headers": {
							"Location": { "$ref": "#/components/headers/Location" }
						}
					},
					"409": { "$ref": "#/components/responses/WorkspaceAlreadyInitialized" }
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"Error": {
				"type": "object",
				"required": [
					"reason"
				],
				"properties": {
					"reason": {
						"type": "string",
						"enum": [
							"bad_request",
							"database_error",
							"internal_server_error",
							"not_found",
							"not_implemented",
							"workspace_already_initialized",
							"workspace_not_initialized",
							"server_config_already_initialized",
							"server_config_not_initialized",
							"first_account_already_initialized",
							"unauthorized",
							"forbidden",
							"unknown"
						],
						"example": "bad_request"
					}
				}
			},
...
			"InvitationTokenType": {
				"type": "string",
				"enum": ["accept", "reject"],
				"example": "accept"
			},
			"AcceptWorkspaceInvitationRequest": {
				"type": "object",
				"required": [
					"nickname",
					"password"
				],
				"properties": {
					"nickname": { "$ref": "#/components/schemas/Nickname" },
					"password": { "$ref": "#/components/schemas/Password" }
				}
			},
...
			"InitWorkspaceRequest": {
				"type": "object",
				"required": [
					"name"
				],
				"properties": {
					"name": { "$ref": "#/components/schemas/WorkspaceName" }
				}
			},
			"InitServerConfigRequest": {
				"type": "object",
				"required": [
					"domain"
				],
				"properties": {
					"domain": {
						"type": "string",
						"example": "crisp.chat"
					}
				}
			},
			"InitFirstAccountRequest": {
				"type": "object",
				"required": [
					"username",
					"password",
					"nickname"
				],
				"properties": {
					"username": { "$ref": "#/components/schemas/Username" },
					"password": { "$ref": "#/components/schemas/Password" },
					"nickname": { "$ref": "#/components/schemas/Nickname" }
				}
			},
			"Workspace": {
				"type": "object",
				"required": [
					"name",
					"accent_color"
				],
				"properties": {
					"name": { "$ref": "#/components/schemas/WorkspaceName" },
					"accent_color": { "$ref": "#/components/schemas/WorkspaceAccentColor" }
				}
			},
			"ServerConfig": {
				"type": "object",
				"required": [
					"domain",
					"message_archive_enabled",
					"message_archive_retention",
					"file_upload_allowed",
					"file_storage_encryption_scheme",
					"file_storage_retention",
					"mfa_required",
					"minimum_tls_version",
					"minimum_cipher_suite",
					"federation_enabled",
					"settings_backup_interval",
					"user_data_backup_interval"
				],
				"properties": {
					"domain": {
						"type": "string",
						"example": "crisp.chat"
					},
					"message_archive_enabled": {
						"type": "boolean",
						"default": true
					},
					"message_archive_retention": {
						"oneOf": [
							{ "$ref": "#/components/schemas/DurationDateLike" },
							{ "$ref": "#/components/schemas/DurationInfinite" }
						],
						"default": "infinite"
					},
					"file_upload_allowed": {
						"type": "boolean",
						"default": true
					},
					"file_storage_encryption_scheme": {
						"type": "string",
						"default": "AES-256"
					},
					"file_storage_retention": {
						"oneOf": [
							{ "$ref": "#/components/schemas/DurationDateLike" },
							{ "$ref": "#/components/schemas/DurationInfinite" }
						],
						"default": "infinite"
					},
					"mfa_required": {
						"title": "Require Multi-Factor Authentication",
						"description": "Whether or not [Multi-Factor Authentication (MFA)](https://en.wikipedia.org/wiki/Multi-factor_authentication) is required for users to log into their account.",
						"type": "boolean",
						"default": true
					},
					"minimum_tls_version": {
						"title": "Minimum SSL/TLS version",
						"description": "The minimum SSL/[TLS](https://fr.wikipedia.org/wiki/Transport_Layer_Security).",
						"type": "string",
						"default": "1.2"
					},
					"minimum_cipher_suite": {
						"title": "Minimum cipher suite (encryption strength)",
						"description": "Accepting lower-security encryption ciphers let older user devices and servers connect to your server, but is also much less secure. A minimum cipher suite of High strength is recommended.",
						"type": "string",
						"enum": ["HIGH_STRENGTH"],
						"default": "HIGH_STRENGTH"
					},
					"federation_enabled": {
						"description": "Whether or not the XMPP server should support interdomain federation. For more information, see [RFC 6120, section 13.14](https://www.rfc-editor.org/rfc/rfc6120#section-13.14).",
						"type": "boolean",
						"default": true
					},
					"settings_backup_interval": {
						"type": "string",
						"format": "duration",
						"default": "P1D"
					},
					"user_data_backup_interval": {
						"type": "string",
						"format": "duration",
						"default": "P1W"
					}
				}
			},
...
			"InviteMemberRequest": {
				"allOf": [
					{ "$ref": "#/components/schemas/InvitationContact" },
					{
						"type": "object",
						"required": [
							"username"
						],
						"properties": {
							"pre_assigned_role": { "$ref": "#/components/schemas/MemberRole" },
							"username": { "$ref": "#/components/schemas/Username" }
						}
					}
				]
			},
...
			"WorkspaceInvitation": {
				"type": "object",
				"required": [
					"invitation_id",
					"created_at",
					"status",
					"jid",
					"pre_assigned_role",
					"contact",
					"accept_token_expires_at"
				],
				"properties": {
					"accept_token_expires_at": { "$ref": "#/components/schemas/DateTime" },
					"contact": { "$ref": "#/components/schemas/InvitationContact" },
					"created_at": { "$ref": "#/components/schemas/DateTime" },
					"invitation_id": { "$ref": "#/components/schemas/InvitationId" },
					"jid": { "$ref": "#/components/schemas/JID" },
					"pre_assigned_role": { "$ref": "#/components/schemas/MemberRole" },
					"status": { "$ref": "#/components/schemas/InvitationStatus" }
				}
			},
			"JID": {
				"summary": "A bare JID, as defined by [RFC 6120, section 2.1](https://www.rfc-editor.org/rfc/rfc6120#section-2.1).",
				"type": "string",
				"format": "jid",
				"example": "remi@prose.org"
			},
			"MemberRole": {
				"summary": "A member's role.",
				"type": "string",
				"enum": ["MEMBER", "ADMIN"]
			},
...
			"InvitationId": {
				"summary": "The unique identifier of an invitation.",
				"type": "integer",
				"format": "int32",
				"example": 42
			},
...
			"Username": {
				"type": "string",
				"summary": "The local part of a JID, as defined by [RFC 6120, section 2.1](https://www.rfc-editor.org/rfc/rfc6120#section-2.1).",
				"example": "remi.bardon"
			},
			"Nickname": {
				"type": "string",
				"summary": "A vCard `NICKNAME` parameter, as defined in [RFC 6350, section 6.2.3](https://datatracker.ietf.org/doc/html/rfc6350#section-6.2.3).",
				"example": "Rémi B."
			},
...
			"Password": {
				"type": "string",
				"format": "password",
				"example": "HsV>dFx^P[m!jM+4W9wf5k:_t{8pSQ*?"
			},
			"DateTime": {
				"type": "string",
				"format": "date-time"
			},
...
			"Member": {
				"type": "object",
				"required": [
					"jid",
					"role"
				],
				"properties": {
					"jid": { "$ref": "#/components/schemas/JID" },
					"role": { "$ref": "#/components/schemas/MemberRole" }
				}
			},
...
		},
		"headers": {
			"Location": {
				"description": "URL to the created resource. See [`Location` on MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Location).",
				"schema": {
					"type": "string",
					"format": "uri"
				},
				"example": "/v1/invitations/42"
			},
...
			"PaginationCurrentPage": {
			  "description": "Current pagination page (non standard header). Starts at `1`.",
				"schema": {
					"type": "integer",
					"format": "int64",
					"minimum": 1
				},
				"example": 1
			},
      "PaginationPageSize": {
			  "description": "Size of a pagination page (non standard header). Starts at `1`.",
				"schema": {
					"type": "integer",
					"format": "int64",
					"minimum": 1,
					"default": 20
				},
				"example": 20
			},
      "PaginationPageCount": {
			  "description": "Total number of pagination pages (non standard header). Warning: `Page-Count` will be `0` with `Current-Page` at `1` if there are no result.",
				"schema": {
					"type": "integer",
					"format": "int64",
					"minimum": 0
				},
				"example": 3
			},
      "PaginationItemCount": {
			  "description": "Total number of items (non standard header).",
				"schema": {
					"type": "integer",
					"format": "int64",
					"minimum": 0
				},
				"example": 42
			}
		},
		"securitySchemes": {
			"BasicAuth": {
				"type": "http",
				"scheme": "basic"
			},
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		},
		"responses": {
			"NoContent": {
				"description": "Success"
			},
			"WorkspaceAlreadyInitialized": {
				"description": "Workspace already initialized",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "workspace_already_initialized"
						}
					}
				}
			},
...
			"ServerConfigAlreadyInitialized": {
				"description": "XMPP server already initialized",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "server_config_already_initialized"
						}
					}
				}
			},
			"ServerConfigNotInitialized": {
				"description": "XMPP server not initialized",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "server_config_not_initialized"
						}
					}
				}
			},
			"FirstAccountAlreadyCreated": {
				"description": "First account already created",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "first_account_already_created"
						}
					}
				}
			},
			"Unauthorized": {
				"description": "Unauthorized",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "unauthorized"
						}
					}
				},
				"headers": {
					"WWW-Authenticate": { "$ref": "#/components/headers/WWW-Authenticate" }
				}
			},
			"Forbidden": {
				"description": "Forbidden",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "forbidden"
						}
					}
				}
			},
			"NotFound": {
				"description": "Object not found",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "not_found"
						}
					}
				}
			},
			"Gone": {
				"description": "Object does not exist",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" },
						"example": {
							"reason": "not_found"
						}
					}
				}
			}
		},
		"parameters": {
			"InvitationId": {
				"description": "Unique identifier of a workspace invitation",
				"name": "invitationId",
				"in": "path",
				"schema": { "$ref": "#/components/schemas/InvitationId" },
				"required": true
			},
			"InvitationIdOrToken": {
				"description": "Unique identifier of a workspace invitation",
				"name": "invitationIdOrToken",
				"in": "path",
				"schema": {
					"oneOf": [
						{ "$ref": "#/components/schemas/InvitationId" },
						{ "$ref": "#/components/schemas/InvitationToken" }
					]
				},
				"required": true
			},
...
			"InvitationAcceptToken": {
				"description": "Invitation accept token",
				"in": "path",
				"name": "token",
				"schema": { "$ref": "#/components/schemas/InvitationAcceptToken" },
				"required": true
			},
...
			"PageNumber": {
				"in": "query",
				"name": "page_number",
				"schema": {
					"type": "integer",
					"format": "int64",
					"default": 1,
					"example": 2
				}
			},
			"PageSize": {
				"in": "query",
				"name": "page_size",
				"schema": {
					"type": "integer",
					"format": "int64",
					"default": 20,
					"example": 50
				}
			},
			"Until": {
				"in": "query",
				"name": "until",
				"schema": { "$ref": "#/components/schemas/DateTime" }
			},
			"MemberId": {
				"name": "memberId",
				"in": "path",
				"required": true,
				"schema": { "$ref": "#/components/schemas/JID" }
			},
...
		}
	}
}